+++ lua/plugins/lsp.lua
@@ -28,11 +28,8 @@ return {
         vim.list_extend(opts.ensure_installed, {
           "bicep-lsp",
           "shellcheck",
-          "flake8", --  python code linter
-          "debugpy", --python debugger
-          "ruff", -- fast python linter, written in Rust.
-          "black", -- python code formatter
-          "isort", --organize python imports
+          "debugpy",
+          "ruff",
           "markdownlint",
           "clang-format",
           "cspell",
@@ -149,11 +146,6 @@ return {
                 workspaceWord = true,
                 callSnippet = "Both",
               },
-              misc = {
-                parameters = {
-                  "--log-level=trace",
-                },
-              },
               diagnostics = {
                 -- enable = false,
                 groupSeverity = {
@@ -197,73 +189,4 @@ return {
       },
     },
   },
-
-  -- NOTE: null-ls.nvim is deprecated and has been replaced by none-ls.nvim
-  -- However, LazyVim now uses conform.nvim for formatting and nvim-lint for linting by default
-  -- which are more modern and better maintained. The configuration below has been disabled.
-  --
-  -- If you want to use none-ls.nvim instead, enable this extra in lazy.lua:
-  -- { import = "lazyvim.plugins.extras.lsp.none-ls" }
-  --
-  -- Then you can configure it here. For now, we're using LazyVim's defaults.
-  --
-  -- Old null-ls configuration (for reference):
-  -- {
-  --   "jose-elias-alvarez/null-ls.nvim",
-  --   opts = function()
-  --     local nls = require("null-ls")
-  --     return {
-  --       root_dir = require("null-ls.utils").root_pattern(".null-ls-root", ".neoconf.json", "Makefile", ".git"),
-  --       sources = {
-  --         nls.builtins.diagnostics.markdownlint,
-  --         nls.builtins.formatting.stylua,
-  --         nls.builtins.diagnostics.selene.with({
-  --           condition = function(utils)
-  --             return utils.root_has_file({ "selene.toml" })
-  --           end,
-  --         }),
-  --         nls.builtins.formatting.isort,
-  --         nls.builtins.formatting.shfmt,
-  --         nls.builtins.formatting.black,
-  --         nls.builtins.diagnostics.flake8,
-  --         nls.builtins.diagnostics.luacheck.with({
-  --           condition = function(utils)
-  --             return utils.root_has_file({ ".luacheckrc" })
-  --           end,
-  --         }),
-  --       },
-  --     }
-  --   end,
-  -- },
-
-  -- {
-  --   "mfussenegger/nvim-dap",
-  --   dependencies = {
-  --     "mfussenegger/nvim-dap-python",
-  --     config = function()
-  --       require("dap-python").setup() -- Use default python
-  --     end,
-  --   },
-  -- },
-  --
-  -- I need to research this a bit more
-  -- inlay hints
-  --   {
-  --     "lvimuser/lsp-inlayhints.nvim",
-  --     event = "LspAttach",
-  --     opts = {},
-  --     config = function(_, opts)
-  --       require("lsp-inlayhints").setup(opts)
-  --       vim.api.nvim_create_autocmd("LspAttach", {
-  --         group = vim.api.nvim_create_augroup("LspAttach_inlayhints", {}),
-  --         callback = function(args)
-  --           if not (args.data and args.data.client_id) then
-  --             return
-  --           end
-  --           local client = vim.lsp.get_client_by_id(args.data.client_id)
-  --           require("lsp-inlayhints").on_attach(client, args.buf)
-  --         end,
-  --       })
-  --     end,
-  --   },
 }